refactor(datagrid): keep the per-cell display path clear of the preference layer
CellDisplayFormatter called itself a pure formatter while its compile
closure was 1076 of the app's 2010 Swift files, reaching PluginManager,
KeychainHelper, MainMenuBuilder and MCPServerManager.

Two edges did it. ValueDisplayFormatService mixed pure value rendering with
storage-backed override resolution, so the grid's per-cell path dragged
ValueDisplayFormatStorage and AppSettingsManager behind it; the rendering
moves to ValueDisplayFormatter, which holds no storage. And DateFormatOption,
the formatting layer's own vocabulary, was declared inside AppSettings.swift,
so every date formatter compiled through the settings model.

The closure is now 46 files.
